v1.8.770

+++ chromium/build/android/pylib/local/device/local_device_environment.py
@@ -2,6 +2,7 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
+import contextlib
 import datetime
 import functools
 import logging
@@ -20,6 +21,8 @@ from devil.utils import file_utils
 from devil.utils import parallelizer
 from pylib import constants
 from pylib.base import environment
+from py_trace_event import trace_event
+from tracing_build import trace2html
 
 
 def _DeviceCachePath(device):
@@ -90,6 +93,7 @@ class LocalDeviceEnvironment(environment.Environment):
     self._skip_clear_data = args.skip_clear_data
     self._target_devices_file = args.target_devices_file
     self._tool_name = args.tool
+    self._trace_output = args.trace_output
 
   #override
   def SetUp(self):
@@ -121,6 +125,8 @@ class LocalDeviceEnvironment(environment.Environment):
 
     @handle_shard_failures_with(on_failure=self.BlacklistDevice)
     def prepare_device(d):
+      d.WaitUntilFullyBooted(timeout=10)
+
       if self._enable_device_cache:
         cache_path = _DeviceCachePath(d)
         if os.path.exists(cache_path):
@@ -141,6 +147,15 @@ class LocalDeviceEnvironment(environment.Environment):
         monitor.Start()
 
     self.parallel_devices.pMap(prepare_device)
+
+  @staticmethod
+  def _JsonToTrace(json_path, html_path, delete_json=True):
+    # First argument is call site.
+    cmd = [__file__, json_path, '--title', 'Android Test Runner Trace',
+           '--output', html_path]
+    trace2html.Main(cmd)
+    if delete_json:
+      os.remove(json_path)
 
   @property
   def blacklist(self):
@@ -176,6 +191,10 @@ class LocalDeviceEnvironment(environment.Environment):
   def tool(self):
     return self._tool_name
 
+  @property
+  def trace_output(self):
+    return self._trace_output
+
   #override
   def TearDown(self):
     if self._devices is None:
@@ -187,9 +206,14 @@ class LocalDeviceEnvironment(environment.Environment):
       # so that an invalid cache can be flushed just by disabling it for one
       # run.
       cache_path = _DeviceCachePath(d)
-      with open(cache_path, 'w') as f:
-        f.write(d.DumpCacheData())
-        logging.info('Wrote device cache: %s', cache_path)
+      if os.path.exists(os.path.dirname(cache_path)):
+        with open(cache_path, 'w') as f:
+          f.write(d.DumpCacheData())
+          logging.info('Wrote device cache: %s', cache_path)
+      else:
+        logging.warning(
+            'Unable to write device cache as %s directory does not exist',
+            os.path.dirname(cache_path))
 
     self.parallel_devices.pMap(tear_down_device)
 
@@ -224,3 +248,24 @@ class LocalDeviceEnvironment(environment.Environment):
     with self._devices_lock:
       self._devices = [d for d in self._devices if str(d) != device_serial]
 
+  def DisableTracing(self):
+    if not trace_event.trace_is_enabled():
+      logging.warning('Tracing is not running.')
+    else:
+      trace_event.trace_disable()
+    self._JsonToTrace(self._trace_output + '.json',
+                      self._trace_output)
+
+  def EnableTracing(self):
+    if trace_event.trace_is_enabled():
+      logging.warning('Tracing is already running.')
+    else:
+      trace_event.trace_enable(self._trace_output + '.json')
+
+  @contextlib.contextmanager
+  def Tracing(self):
+    try:
+      self.EnableTracing()
+      yield
+    finally:
+      self.DisableTracing()
